Unify logging into SullaLogger with file-based topic + conversation logging
Merges ConversationLogger and the Electron Log-based approach into a single
SullaLogger service that writes all logs to ~/sulla/logs/. Topic loggers
(websocket, chat, persona, dispatcher, frontend-graph) replace console.log
in WebSocket, chat, and message dispatch code so infrastructure events are
persisted to disk for debugging stale-socket and lost-message issues.
